[FEAT] 로그아웃 api 개발 완료(#S08P22D208-296), [FEAT] 유저 지갑주소 저장 api 개발(#S08P22D208-297)

See merge request s08-blockchain-contract-sub2/S08P22D208!54

+++ Server/module-core/src/main/java/com/example/kkaddak/core/repository/FollowRepository.java
@@ -13,6 +13,6 @@ public interface FollowRepository extends JpaRepository<Follow, Integer> {
     List<Follow> findByFollowing(Member following);
     Optional<Follow> findByFollowerAndFollowing(Member follower, Member following);
 
-    Long countByFollower(Member follower);
-    Long countByFollowing(Member following);
+    int countByFollower(Member follower);
+    int countByFollowing(Member following);
 }
